PC1/microservices/command-parser/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
kafka-python==2.0.2
pydantic==2.5.0
spacy==3.7.2
python-multipart==0.0.6
